refactor: refactor pull command, move issue pull and compute to pkg/repo
