+++ cmd/cmd_db.go
@@ -69,34 +69,3 @@ func dbDump(opts *dbOptions) error {
 	fmt.Println(string(out))
 	return nil
 }
-
-func loadIssues(targets []string) (repo.Issues, error) {
-	query := db.Model(repo.Issue{}).Order("created_at")
-	if len(targets) > 0 {
-		return nil, fmt.Errorf("not implemented")
-		// query = query.Where("repo_url IN (?)", canonicalTargets(targets))
-		// OR WHERE parents IN ....
-		// etc
-	}
-
-	perPage := 100
-	var issues []*repo.Issue
-	for page := 0; ; page++ {
-		var newIssues []*repo.Issue
-		if err := query.Limit(perPage).Offset(perPage * page).Find(&newIssues).Error; err != nil {
-			return nil, err
-		}
-		issues = append(issues, newIssues...)
-		if len(newIssues) < perPage {
-			break
-		}
-	}
-
-	for _, issue := range issues {
-		issue.PostLoad()
-	}
-
-	return repo.Issues(issues), nil
-}
-
-// FIXME: try to use gorm hooks to auto preload/postload items
