+++ cmd/cmd_pull.go
@@ -2,15 +2,14 @@ package main
 
 import (
 	"encoding/json"
-	"moul.io/depviz/pkg/repo"
 	"os"
-	"sync"
 
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 	"go.uber.org/zap"
+	"moul.io/depviz/pkg/repo"
 )
 
 type pullOptions struct {
@@ -61,51 +60,14 @@ func (cmd *pullCommand) NewCobraCommand(dc map[string]DepvizCommand) *cobra.Comm
 }
 
 func pullAndCompute(opts *pullOptions) error {
+	zap.L().Debug("pull", zap.Stringer("opts", *opts))
 	if os.Getenv("DEPVIZ_NOPULL") != "1" {
-		if err := pull(opts); err != nil {
+		if err := repo.PullAndCompute(opts.GithubToken, opts.GitlabToken, db, opts.Targets); err != nil {
 			return errors.Wrap(err, "failed to pull")
 		}
-	}
-	if err := compute(opts); err != nil {
-		return errors.Wrap(err, "failed to compute")
-	}
-	return nil
-}
-
-func pull(opts *pullOptions) error {
-	// FIXME: handle the special '@me' target
-	zap.L().Debug("pull", zap.Stringer("opts", *opts))
-
-	var (
-		wg        sync.WaitGroup
-		allIssues []*repo.Issue
-		out       = make(chan []*repo.Issue, 100)
-	)
-
-	targets := opts.Targets.UniqueProjects()
-
-	// parallel fetches
-	wg.Add(len(targets))
-	for _, target := range targets {
-		switch target.Driver() {
-		case repo.GithubDriver:
-			go repo.GithubPull(target, &wg, opts.GithubToken, db, out)
-		case repo.GitlabDriver:
-			go repo.GitlabPull(target, &wg, opts.GitlabToken, db, out)
-		default:
-			panic("should not happen")
-		}
-	}
-	wg.Wait()
-	close(out)
-	for issues := range out {
-		allIssues = append(allIssues, issues...)
-	}
-
-	// save
-	for _, issue := range allIssues {
-		if err := db.Save(issue).Error; err != nil {
-			return err
+	} else {
+		if err := repo.Compute(db); err != nil {
+			return errors.Wrap(err, "failed to compute")
 		}
 	}
 	return nil
