+++ pkg/repo/target.go
@@ -154,17 +154,3 @@ func (t Targets) UniqueProjects() Targets {
 	}
 	return filtered
 }
-
-func uniqueStrings(input []string) []string {
-	u := make([]string, 0, len(input))
-	m := make(map[string]bool)
-
-	for _, val := range input {
-		if _, ok := m[val]; !ok {
-			m[val] = true
-			u = append(u, val)
-		}
-	}
-
-	return u
-}
